ARMCC fiixes from Mike Kasprzak

+++ intrusive_list.h
@@ -90,7 +90,7 @@ private:
 class intrusive_list_base
 {
 public:
-	typedef size_t	size_type;
+	typedef int	size_type;
 
 	void pop_back()
 	{
